feat: added preferences fragment.
