+++ app/build.gradle
@@ -19,6 +19,7 @@ android {
     }
     testOptions {
         unitTests.returnDefaultValues = true
+        unitTests.includeAndroidResources = true
     }
 }
 
@@ -26,12 +27,13 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(':msglibrary')
     implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.preference:preference:1.1.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'com.google.android.material:material:1.1.0'
     implementation 'com.google.android.gms:play-services-location:17.0.0'
     implementation 'androidx.room:room-runtime:2.2.3'
     implementation 'com.github.rehacktive:waspdb:1.1.1'
-    implementation "androidx.fragment:fragment:1.2.1"
+    implementation 'androidx.fragment:fragment:1.2.1'
 
 
     testImplementation 'junit:junit:4.13'
