+++ app/src/androidTest/java/ingsw/group1/findmyphone/AndroidTestUtils.java
@@ -1,0 +1,36 @@
+package ingsw.group1.findmyphone;
+
+import android.view.View;
+
+import androidx.test.espresso.UiController;
+import androidx.test.espresso.ViewAction;
+
+import org.hamcrest.Matcher;
+
+import static androidx.test.espresso.matcher.ViewMatchers.isRoot;
+
+/**
+ * Just a helper class containing methods used when displaying preview tests.
+ *
+ * @author Riccardo De Zen.
+ */
+public class AndroidTestUtils {
+    public static ViewAction waitFor(final long millis) {
+        return new ViewAction() {
+            @Override
+            public Matcher<View> getConstraints() {
+                return isRoot();
+            }
+
+            @Override
+            public String getDescription() {
+                return "Wait for " + millis + " milliseconds.";
+            }
+
+            @Override
+            public void perform(UiController uiController, final View view) {
+                uiController.loopMainThreadForAtLeast(millis);
+            }
+        };
+    }
+}
