+++ app/src/main/java/ingsw/group1/findmyphone/cryptography/PasswordManager.java
@@ -3,7 +3,10 @@ package ingsw.group1.findmyphone.cryptography;
 import android.content.Context;
 import android.content.SharedPreferences;
 
-import static android.content.Context.MODE_PRIVATE;
+import androidx.annotation.NonNull;
+import androidx.preference.PreferenceManager;
+
+import ingsw.group1.findmyphone.R;
 
 /**
  * Class used to store a password.
@@ -11,18 +14,19 @@ import static android.content.Context.MODE_PRIVATE;
  * @author Pardeep Kumar
  */
 public class PasswordManager {
-    private static final String PASSWORD_FILE_NAME = "ingsw.group1.findmyphone.password";
-    private static final String PASSWORD = "password";
-    private static final String DEFAULT_PASSWORD = "ASkj26%&/chfd";
+    private static final String DEFAULT_PASSWORD = "null";
     private SharedPreferences sharedPreferences;
+    private String passwordKey;
 
     /**
-     * Constructor of the passwordManager object, it requires a context to create a sharedPreference object.
+     * Constructor of the passwordManager object, it requires a context to create a
+     * sharedPreference object.
      *
      * @param context The context used to create the sharedPreference.
      */
     public PasswordManager(Context context) {
-        sharedPreferences = context.getSharedPreferences(PASSWORD_FILE_NAME, MODE_PRIVATE);
+        sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context);
+        passwordKey = context.getResources().getString(R.string.password_key);
     }
 
     /**
@@ -32,7 +36,7 @@ public class PasswordManager {
      */
     public void storePassword(String newPassword) {
         SharedPreferences.Editor editor = sharedPreferences.edit();
-        editor.putString(PASSWORD, newPassword);
+        editor.putString(passwordKey, newPassword);
         editor.apply();
     }
 
@@ -41,8 +45,9 @@ public class PasswordManager {
      *
      * @return The string containing the password.
      */
+    @NonNull
     public String retrievePassword() {
-        return sharedPreferences.getString(PASSWORD, DEFAULT_PASSWORD);
+        return sharedPreferences.getString(passwordKey, DEFAULT_PASSWORD);
     }
 
 }
