+++ app/src/main/res/xml/app_preferences.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <EditTextPreference
+        android:inputType="textPassword"
+        app:key="@string/password_key"
+        app:summary="@string/password_preference_text"
+        app:title="@string/password_title" />
+
+</PreferenceScreen>
